refactor: complete refactoring project structure
