+++ src/components/ui/button.tsx
@@ -19,6 +19,7 @@ const buttonVariants = cva(
           'bg-secondary text-secondary-foreground shadow-xs hover:bg-secondary/80',
         ghost:
           'hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50',
+        invisible: 'bg-transparent hover:bg-transparent',
         link: 'text-primary underline-offset-4 hover:underline',
       },
       size: {
