+++ src/components/widgets/language-switcher/language-switcher.tsx
