+++ src/features/auth/ui/sign-in-form.tsx
@@ -74,12 +74,12 @@ export default function SignInForm() {
 
   return (
     <div className="min-h-200 flex items-center justify-center bg-gradient-to-br from-blue-50 to-indigo-100 px-4 py-12">
-      <Card className="w-full max-w-md shadow-xl border-0">
+      <Card className="w-full max-w-md shadow-2xl shadow-primary/10 border-border/50">
         <CardHeader className="space-y-1 text-center">
-          <CardTitle className="text-2xl font-bold text-gray-900">
+          <CardTitle className="text-2xl font-bold text-foreground">
             {t('signInTitle')}
           </CardTitle>
-          <CardDescription className="text-gray-600">
+          <CardDescription className="text-muted-foreground">
             {t('signInDescription')}
           </CardDescription>
         </CardHeader>
@@ -100,14 +100,13 @@ export default function SignInForm() {
                 name="email"
                 render={({ field }) => (
                   <FormItem>
-                    <FormLabel className="text-gray-700 font-medium">
+                    <FormLabel className="text-foreground font-medium">
                       {t('emailLabel')}
                     </FormLabel>
                     <FormControl>
                       <Input
                         placeholder="email@example.com"
                         type="email"
-                        className="border-gray-300 focus:border-blue-500 focus:ring-blue-500"
                         {...field}
                       />
                     </FormControl>
@@ -121,7 +120,7 @@ export default function SignInForm() {
                 name="password"
                 render={({ field }) => (
                   <FormItem>
-                    <FormLabel className="text-gray-700 font-medium">
+                    <FormLabel className="text-foreground font-medium">
                       {t('passwordLabel')}
                     </FormLabel>
                     <FormControl>
@@ -129,12 +128,14 @@ export default function SignInForm() {
                         <Input
                           placeholder={t('passwordPlaceholder')}
                           type={showPassword ? 'text' : 'password'}
-                          className="border-gray-300 focus:border-blue-500 focus:ring-blue-500 pr-10"
+                          className="pr-10"
                           {...field}
                         />
-                        <button
+                        <Button
                           type="button"
-                          className="absolute right-3 top-1/2 transform -translate-y-1/2 text-gray-500 hover:text-gray-700 transition-colors"
+                          variant="ghost"
+                          size="icon"
+                          className="absolute right-1 top-1/2 h-7 w-7 -translate-y-1/2 text-muted-foreground hover:text-foreground transition-colors"
                           onClick={() => setShowPassword(!showPassword)}
                         >
                           {showPassword ? (
@@ -142,7 +143,7 @@ export default function SignInForm() {
                           ) : (
                             <Eye size={18} className="text-gray-400" />
                           )}
-                        </button>
+                        </Button>
                       </div>
                     </FormControl>
                     <FormMessage className="text-destructive" />
@@ -150,24 +151,21 @@ export default function SignInForm() {
                 )}
               />
 
-              <Button
-                type="submit"
-                disabled={pending}
-                className="w-full bg-blue-600 hover:bg-blue-700 text-white font-medium py-3 px-4 rounded-md transition-all duration-300"
-              >
+              <Button type="submit" disabled={pending} className="w-full">
                 {pending ? t('signingIn') : t('signInButton')}
               </Button>
 
               <div className="text-center">
-                <p className="text-sm text-gray-600">
+                <p className="text-sm text-muted-foreground">
                   {t('noAccount')}{' '}
-                  <button
+                  <Button
                     type="button"
+                    variant="link"
+                    className="p-0 h-auto font-medium text-primary"
                     onClick={() => router.push('/auth/signUp')}
-                    className="text-blue-600 hover:text-blue-700 font-medium underline transition-colors"
                   >
                     {t('signUpLink')}
-                  </button>
+                  </Button>
                 </p>
               </div>
             </form>
