--- src/screens/auth/page.client.tsx
@@ -1,7 +1,0 @@
-'use client';
-
-import { SignInForm, SignUpForm } from '@/features/auth';
-
-export function AuthScreen({ authType }: { authType: 'signIn' | 'signUp' }) {
-  return authType === 'signIn' ? <SignInForm /> : <SignUpForm />;
-}
