--- src/screens/history/page.server.tsx
@@ -1,17 +1,0 @@
-import { getHistory } from '@/features/history';
-import { HistoryList } from '@/features/history';
-import { getTranslations } from 'next-intl/server';
-
-export const dynamic = 'force-dynamic';
-
-export default async function HistoryScreen() {
-  const history = await getHistory();
-  const t = await getTranslations('History');
-
-  return (
-    <div className="p-4 space-y-6 max-w-4xl mx-auto">
-      <h1 className="text-3xl font-bold">{t('title')}</h1>
-      <HistoryList history={history} />
-    </div>
-  );
-}
