+++ vitest.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
       include: ['src/**/*.{js,jsx,ts,tsx}'],
       exclude: [
         'src/components/ui/**',
-        'src/i18n/**',
+        'src/core/i18n/**',
         'src/middleware.ts',
         'src/**/*.test.{js,jsx,ts,tsx}',
         'src/index.{js,jsx,ts,tsx}',
